+++ packages/nimble-components/.storybook/preview.js
@@ -15,12 +15,20 @@ export const parameters = {
     options: {
         storySort: {
             method: 'alphabetical',
+            // Items within arrays show the sort order for children of the category above
+            // https://storybook.js.org/docs/react/writing-stories/naming-components-and-hierarchy#sorting-stories
             order: [
                 'Getting Started',
                 'Components',
+                ['Status Table'],
                 'Incubating',
+                ['Docs'],
                 'Tokens',
-                'Tests'
+                ['Docs'],
+                'Tests',
+                ['Docs'],
+                'Internal',
+                ['Docs']
             ]
         }
     },